% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_dpd_stats.R
\name{ffp_dpd_atkinson_drm}
\alias{ffp_dpd_atkinson_drm}
\title{Compute Atkinson (1970) Inequality Index}
\usage{
ffp_dpd_atkinson_drm(ar_data, ar_prob_data, fl_rho)
}
\arguments{
\item{ar_data}{array sorted array values}

\item{ar_prob_data}{array probability mass for each element along \code{ar_data}, sums to 1}

\item{fl_rho}{float inequality aversion parameter fl_rho = 1 for planner
without inequality aversion. fl_rho = -infinity for fully inequality averse.}
}
\value{
A scalar value for the Atkinson index, between 0 and < 1
}
\description{
Atkinson inequality index
}
\examples{
ffp_dpd_atkinson_drm(c(1,2,3), c(1/3, 1/3, 1/3), 0.65)
ffp_dpd_atkinson_drm(c(1,2,3), c(1/3, 1/3, 1/3), 0.25)
ffp_dpd_atkinson_drm(c(1,2,3), c(1/3, 1/3, 1/3), 0.15)
ffp_dpd_atkinson_drm(c(1e-5, 1e-5, 1e5), c(1/3, 1/3, 1/3), 0.65)
ffp_dpd_atkinson_drm(c(1e-5, 1e-5, 1e5), c(1/3, 1/3, 1/3), 0.25)
ffp_dpd_atkinson_drm(c(1e-5, 1e-5, 1e5), c(1/3, 1/3, 1/3), 0.15)
}
\references{
\href{https://fanwangecon.github.io/R4Econ/math/func_ineq/htmlpdfr/fs_gini_disc.html}{fs_gini_disc}
}
\seealso{
\code{\link[=ffp_dpd_gini_drm]{ffp_dpd_gini_drm()}} for the Gini coefficient and \code{\link[=ffp_dpd_cov_drm]{ffp_dpd_cov_drm()}}
for the coefficient of variation.
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
