% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_dpd_stats.R
\name{ffp_dpd_cov_drm}
\alias{ffp_dpd_cov_drm}
\title{Compute Coefficient of Variation given discrete distribution}
\usage{
ffp_dpd_cov_drm(ar_data, ar_prob_data, verbose = FALSE)
}
\arguments{
\item{ar_data}{array array values}

\item{ar_prob_data}{array probability mass for each element along \code{ar_data}, sums to 1}
}
\value{
A scalar value for the the coefficient of variation, always positive
}
\description{
Coefficient of variation.
}
\examples{
ffp_dpd_cov_drm(c(1,2,3), c(1/3, 1/3, 1/3))
ffp_dpd_cov_drm(c(1e-5, 1e-5, 1e5), c(1/3, 1/3, 1/3))
}
\references{
\href{https://fanwangecon.github.io/R4Econ/math/func_ineq/htmlpdfr/fs_gini_disc.html}{fs_gini_disc}
}
\seealso{
\code{\link[=ffp_dpd_gini_drm]{ffp_dpd_gini_drm()}} for the Gini coefficient and \code{\link[=ffp_dpd_atkinson_drm]{ffp_dpd_atkinson_drm()}}
for the Atkinson Index.
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
