% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_dpd_stats.R
\name{ffp_dpd_gini_drm}
\alias{ffp_dpd_gini_drm}
\title{Compute GINI given discrete distribution}
\usage{
ffp_dpd_gini_drm(ar_data, ar_prob_data)
}
\arguments{
\item{ar_data}{array sorted array values low to high}

\item{ar_prob_data}{array probability mass for each element along \code{ar_data}, sums to 1}
}
\value{
A scalar value for the GINI coefficient, between 0 and < 1
}
\description{
Discrete distribution-specific GINI.
}
\examples{
ffp_dpd_gini_drm(c(1,2,3), c(1/3, 1/3, 1/3))
ffp_dpd_gini_drm(c(1e-5, 1e-5, 1e5), c(1/3, 1/3, 1/3))
}
\references{
\href{https://fanwangecon.github.io/R4Econ/math/func_ineq/htmlpdfr/fs_gini_disc.html}{fs_gini_disc}
}
\seealso{
\code{\link[=ffp_dpd_cov_drm]{ffp_dpd_cov_drm()}} for coefficient of variation and \code{\link[=ffp_dpd_atkinson_drm]{ffp_dpd_atkinson_drm()}}
for the Atkinson Index.
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
